Added documentation info for OAuth2
This adds the necessary environment variables to the list and adds a link to the
Googgle OAuth2 documentation. We're relying on the python-social-auth module to
actually perform the authorization steps for us, but the OAuth2 documentation
tells us about actually getting the client id and secret. It's a pretty simple
process -- just go into the Google development console and grab the client id
and secret from there.